Big Changes
 1. Added a releaseIntervalMs to channel state to rate limit messages
 2. Used a BIGSERIAL instead of an UUID for table IDs for better
    logical ordering (esp within single transactions)
 3. Renamed dequeueId -> dequeueNonce
 4. Tidied up some orphaned fields

+++ src/migration/03-function-wake.ts
@@ -12,7 +12,7 @@ export const migrationFunctionWake = {
         return [
             sql`
                 CREATE FUNCTION ${ref(params.schema)}."wake" (
-                    p_delay_ms BIGINT
+                    p_delay_ms INTEGER
                 ) RETURNS VOID AS $$
                 BEGIN
                     IF ${value(params.useWake)} THEN
